requirements.txt
flask
flask-cors
pydub
soundfile
matplotlib
numpy
gunicorn
ffmpeg-python
